# Enable AVX-512 for string/span Equals/StartsWith